Dragons Renewal! RW 1.6 Update!
Added over 60 new graphics for dragons to appear with.

Set up the code for new breath abilities that will use Biotech code as its far more flexible and stable.
(Planning for that to be the next major update)

Lots of Misc cleaning up of code because some of it is over 5 years old.